feat(dashboard): add link to parent wfRun
