+++ dashboard-new/src/app/(authenticated)/wfSpec/[...props]/components/WfRuns.tsx
@@ -1,6 +1,7 @@
 'use client'
 import { SearchFooter } from '@/app/(authenticated)/components/SearchFooter'
 import { SEARCH_DEFAULT_LIMIT, TIME_RANGES, TimeRange } from '@/app/constants'
+import { concatWfRunIds } from '@/app/utils/wfRun'
 import { useWhoAmI } from '@/contexts/WhoAmIContext'
 import { ArrowPathIcon } from '@heroicons/react/16/solid'
 import { useInfiniteQuery } from '@tanstack/react-query'
@@ -12,7 +13,6 @@ import { useSearchParams } from 'next/navigation'
 import { FC, Fragment, useMemo, useState } from 'react'
 import { searchWfRun } from '../actions/searchWfRun'
 import { WfRunsHeader } from './WfRunsHeader'
-import { WfRunId } from 'littlehorse-client/dist/proto/object_id'
 
 type Props = Pick<WfSpec, 'id'>
 export const WfRuns: FC<Props> = ({ id }) => {
@@ -63,9 +63,9 @@ export const WfRuns: FC<Props> = ({ id }) => {
         <div className="flex min-h-[360px] flex-col gap-4">
           {data?.pages.map((page, i) => (
             <Fragment key={i}>
-              {page.results.map((wfRunId) => (
+              {page.results.map(wfRunId => (
                 <div key={wfRunId.id}>
-                  <Link className="py-2 text-blue-500 hover:underline" href={`/wfRun/${concatIds(wfRunId)}`}>
+                  <Link className="py-2 text-blue-500 hover:underline" href={`/wfRun/${concatWfRunIds(wfRunId)}`}>
                     {wfRunId.id}
                   </Link>
                 </div>
@@ -82,14 +82,4 @@ export const WfRuns: FC<Props> = ({ id }) => {
 const getStatus = (status: string | null) => {
   if (!status) return undefined
   return lHStatusFromJSON(status)
-}
-
-const concatIds = (wfRunId: WfRunId) => {
-  const ids = []
-  let current: WfRunId | undefined = wfRunId
-  while (current) {
-    ids.push(current.id)
-    current = current.parentWfRunId
-  }
-  return ids.reverse().join('/')
 }
